services/ark-api-a2a/pyproject.toml
[project]
name = "a2agw"
version = "0.1.15"
description = "Agent-to-Agent Gateway for ARK ecosystem"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
    "a2a-sdk>=0.2.12",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "a2a-sdk[http-server]",
    "coverage>=7.10.4",
    "ark-sdk",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true


[tool.coverage.run]
source = ["a2agw"]
omit = ["tests/*"]
data_file = "coverage/.coverage"

[tool.coverage.html]
directory = "coverage"

[tool.coverage.lcov]
output = "coverage/lcov.info"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.uv.sources]
ark-sdk = { path = "../../out/ark-sdk/py-sdk/dist/ark_sdk-0.1.25-py3-none-any.whl" }
